Moved classes into folders for organization
