+++ BCHA-Formulary/app/src/main/java/com/lowermainlandpharmacyservices/lmpsformulary/Model/BrandDrug.java
@@ -1,4 +1,4 @@
-package com.lowermainlandpharmacyservices.lmpsformulary;
+package com.lowermainlandpharmacyservices.lmpsformulary.Model;
 
 import java.util.ArrayList;
