--- BCHA-Formulary/app/src/main/java/com/lowermainlandpharmacyservices/lmpsformulary/Utilities.java
@@ -1,17 +1,0 @@
-package com.lowermainlandpharmacyservices.lmpsformulary;
-
-public final class Utilities {
-
-	// Password SECTION
-	static String password = "bettercare";
-	static String authorizedUser = "authorization";
-
-	// Keys passing data SECTION
-	static String EXTRA_INFO = "com.lowermainlandpharmacyservices.MainActivity.SEARCHINPUT";
-	static String EXTRA_STRENGTHS = "com.lowermainlandpharmacyservices.MainActivity.STRENGTHS";
-	static String EXTRA_GENERICNAME = "com.lowermainlandpharmacyservices.MainActivity.GENERICNAME";
-	static String EXTRA_BRANDNAME = "com.lowermainlandpharmacyservices.MainActivity.BRANDNAME";
-	static String EXTRA_RESTRICTIONS = "com.lowermainlandpharmacyservices.MainActivity.RESTRICTIONS";
-	static String EXTRA_EXCLUDED_REASON = "com.lowermainlandpharmacyservices.MainActivity.EXCLUDED_REASON";
-	static String EXTRA_TYPE = "com.lowermainlandpharmacyservices.MainActivity.TYPE";
-}
