+++ BCHA-Formulary/app/src/main/java/com/lowermainlandpharmacyservices/lmpsformulary/Utilities/Utilities.java
@@ -1,0 +1,17 @@
+package com.lowermainlandpharmacyservices.lmpsformulary.Utilities;
+
+public final class Utilities {
+
+	// Password SECTION
+	public static String password = "bettercare";
+	public static String authorizedUser = "authorization";
+
+	// Keys passing data SECTION
+	public static String EXTRA_INFO = "com.lowermainlandpharmacyservices.MainActivity.SEARCHINPUT";
+	public static String EXTRA_STRENGTHS = "com.lowermainlandpharmacyservices.MainActivity.STRENGTHS";
+	public static String EXTRA_GENERICNAME = "com.lowermainlandpharmacyservices.MainActivity.GENERICNAME";
+	public static String EXTRA_BRANDNAME = "com.lowermainlandpharmacyservices.MainActivity.BRANDNAME";
+	public static String EXTRA_RESTRICTIONS = "com.lowermainlandpharmacyservices.MainActivity.RESTRICTIONS";
+	public static String EXTRA_EXCLUDED_REASON = "com.lowermainlandpharmacyservices.MainActivity.EXCLUDED_REASON";
+	public static String EXTRA_TYPE = "TYPE";
+}
