+++ BCHA-Formulary/app/src/main/res/layout/activity_display_formulary_result.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     android:background="#ffffff"
     android:orientation="vertical"
-    tools:context="com.lowermainlandpharmacyservices.lmpsformulary.DisplayFormularyResult" >
+    tools:context=".Activity.DisplayFormularyResult" >
 
     <TextView
         android:id="@+id/formulary_drug_genericname"
