+++ BCHA-Formulary/app/src/main/res/menu/display_excluded_result.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.lowermainlandpharmacyservices.lmpsformulary.DisplayExcludedResult" >
+    tools:context=".Activity.DisplayExcludedResult" >
 
     <item
         android:id="@+id/action_settings"
