+++ BCHA-Formulary/app/src/main/res/menu/display_formulary_result.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.lowermainlandpharmacyservices.lmpsformulary.DisplayFormularyResult" >
+    tools:context=".Activity.DisplayFormularyResult" >
 
     <item
         android:id="@+id/action_settings"
