+++ BCHA-Formulary/app/src/main/res/menu/display_other_result.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.lowermainlandpharmacyservices.lmpsformulary.DisplayOtherResult" >
+    tools:context=".Activity.DisplayOtherResult" >
 
     <item
         android:id="@+id/action_settings"
