+++ BCHA-Formulary/app/src/main/res/menu/main.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.lowermainlandpharmacyservices.lmpsformulary.MainActivity" >
+    tools:context=".Activity.MainActivity" >
 
     <item
         android:id="@+id/action_settings"
