+++ BCHA-Formulary/app/src/main/res/menu/splash_screen.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.lowermainlandpharmacyservices.lmpsformulary.SplashScreenActivity" >
+    tools:context=".Activity.SplashScreenActivity" >
 
     <item
         android:id="@+id/action_settings"
